requirements.txt
snowflake-connector-python[pandas]
snowflake-snowpark-python
faker
plotly
streamlit
python-dotenv
pytest
black
flake8
